+++ src/Element.cc
@@ -31,6 +31,7 @@ Element::Element()
 {
   this->dataPtr->copyChildren = false;
   this->dataPtr->referenceSDF = "";
+  this->dataPtr->explicitlySetInFile = true;
 }
 
 /////////////////////////////////////////////////
@@ -91,6 +92,25 @@ const std::string &Element::GetRequired() const
 void Element::SetCopyChildren(bool _value)
 {
   this->dataPtr->copyChildren = _value;
+}
+
+/////////////////////////////////////////////////
+void Element::SetExplicitlySetInFile(const bool _value)
+{
+  this->dataPtr->explicitlySetInFile = _value;
+
+  ElementPtr_V::const_iterator eiter;
+  for (eiter = this->dataPtr->elements.begin();
+       eiter != this->dataPtr->elements.end(); ++eiter)
+  {
+    (*eiter)->SetExplicitlySetInFile(_value);
+  }
+}
+
+/////////////////////////////////////////////////
+bool Element::GetExplicitlySetInFile() const
+{
+  return this->dataPtr->explicitlySetInFile;
 }
 
 /////////////////////////////////////////////////
@@ -164,12 +184,12 @@ ElementPtr Element::Clone() const
   clone->dataPtr->name = this->dataPtr->name;
   clone->dataPtr->required = this->dataPtr->required;
   clone->dataPtr->copyChildren = this->dataPtr->copyChildren;
-  clone->dataPtr->includeFilename = this->dataPtr->includeFilename;
   clone->dataPtr->referenceSDF = this->dataPtr->referenceSDF;
   clone->dataPtr->path = this->dataPtr->path;
   clone->dataPtr->lineNumber = this->dataPtr->lineNumber;
   clone->dataPtr->xmlPath = this->dataPtr->xmlPath;
   clone->dataPtr->originalVersion = this->dataPtr->originalVersion;
+  clone->dataPtr->explicitlySetInFile = this->dataPtr->explicitlySetInFile;
 
   Param_V::const_iterator aiter;
   for (aiter = this->dataPtr->attributes.begin();
@@ -212,12 +232,12 @@ void Element::Copy(const ElementPtr _elem)
   this->dataPtr->description = _elem->GetDescription();
   this->dataPtr->required = _elem->GetRequired();
   this->dataPtr->copyChildren = _elem->GetCopyChildren();
-  this->dataPtr->includeFilename = _elem->dataPtr->includeFilename;
   this->dataPtr->referenceSDF = _elem->ReferenceSDF();
   this->dataPtr->originalVersion = _elem->OriginalVersion();
   this->dataPtr->path = _elem->FilePath();
   this->dataPtr->lineNumber = _elem->LineNumber();
   this->dataPtr->xmlPath = _elem->XmlPath();
+  this->dataPtr->explicitlySetInFile = _elem->GetExplicitlySetInFile();
 
   for (Param_V::iterator iter = _elem->dataPtr->attributes.begin();
        iter != _elem->dataPtr->attributes.end(); ++iter)
@@ -465,47 +485,59 @@ void Element::PrintDocLeftPane(std::string &_html, int _spacing,
 }
 
 void Element::PrintValuesImpl(const std::string &_prefix,
+                              bool _includeDefaultElements,
+                              bool _includeDefaultAttributes,
                               std::ostringstream &_out) const
 {
-  _out << _prefix << "<" << this->dataPtr->name;
+  if (this->GetExplicitlySetInFile() || _includeDefaultElements)
+  {
+    _out << _prefix << "<" << this->dataPtr->name;
 
-  Param_V::const_iterator aiter;
-  for (aiter = this->dataPtr->attributes.begin();
-       aiter != this->dataPtr->attributes.end(); ++aiter)
-  {
-    // Only print attribute values if they were set
-    // TODO(anyone): GetRequired is added here to support up-conversions where a
-    // new required attribute with a default value is added. We would have
-    // better separation of concerns if the conversion process set the required
-    // attributes with their default values.
-    if ((*aiter)->GetSet() || (*aiter)->GetRequired())
+    Param_V::const_iterator aiter;
+    for (aiter = this->dataPtr->attributes.begin();
+         aiter != this->dataPtr->attributes.end(); ++aiter)
     {
-      _out << " " << (*aiter)->GetKey() << "='"
-           << (*aiter)->GetAsString() << "'";
+      // Only print attribute values if they were set
+      // TODO(anyone): GetRequired is added here to support up-conversions where
+      // a new required attribute with a default value is added. We would have
+      // better separation of concerns if the conversion process set the
+      // required attributes with their default values.
+      if ((*aiter)->GetSet() || (*aiter)->GetRequired() ||
+          _includeDefaultAttributes)
+      {
+        _out << " " << (*aiter)->GetKey() << "='"
+             << (*aiter)->GetAsString() << "'";
+      }
     }
-  }
 
-  if (this->dataPtr->elements.size() > 0)
-  {
-    _out << ">\n";
-    ElementPtr_V::const_iterator eiter;
-    for (eiter = this->dataPtr->elements.begin();
-         eiter != this->dataPtr->elements.end(); ++eiter)
+    if (this->dataPtr->elements.size() > 0)
     {
-      (*eiter)->ToString(_prefix + "  ", _out);
-    }
-    _out << _prefix << "</" << this->dataPtr->name << ">\n";
-  }
-  else
-  {
-    if (this->dataPtr->value)
-    {
-      _out << ">" << this->dataPtr->value->GetAsString()
-           << "</" << this->dataPtr->name << ">\n";
+      _out << ">\n";
+      ElementPtr_V::const_iterator eiter;
+      for (eiter = this->dataPtr->elements.begin();
+           eiter != this->dataPtr->elements.end(); ++eiter)
+      {
+        if ((*eiter)->GetExplicitlySetInFile() || _includeDefaultElements)
+        {
+          (*eiter)->ToString(_prefix + "  ",
+                             _includeDefaultElements,
+                             _includeDefaultAttributes,
+                             _out);
+        }
+      }
+      _out << _prefix << "</" << this->dataPtr->name << ">\n";
     }
     else
     {
-      _out << "/>\n";
+      if (this->dataPtr->value)
+      {
+        _out << ">" << this->dataPtr->value->GetAsString()
+             << "</" << this->dataPtr->name << ">\n";
+      }
+      else
+      {
+        _out << "/>\n";
+      }
     }
   }
 }
@@ -514,7 +546,20 @@ void Element::PrintValuesImpl(const std::string &_prefix,
 void Element::PrintValues(std::string _prefix) const
 {
   std::ostringstream ss;
-  PrintValuesImpl(_prefix, ss);
+  PrintValuesImpl(_prefix, true, false, ss);
+  std::cout << ss.str();
+}
+
+/////////////////////////////////////////////////
+void Element::PrintValues(const std::string &_prefix,
+                          bool _includeDefaultElements,
+                          bool _includeDefaultAttributes) const
+{
+  std::ostringstream ss;
+  PrintValuesImpl(_prefix,
+                  _includeDefaultElements,
+                  _includeDefaultAttributes,
+                  ss);
   std::cout << ss.str();
 }
 
@@ -522,23 +567,33 @@ void Element::PrintValues(std::string _prefix) const
 std::string Element::ToString(const std::string &_prefix) const
 {
   std::ostringstream out;
-  this->ToString(_prefix, out);
+  this->ToString(_prefix, true, false, out);
+  return out.str();
+}
+
+/////////////////////////////////////////////////
+std::string Element::ToString(const std::string &_prefix,
+                              bool _includeDefaultElements,
+                              bool _includeDefaultAttributes) const
+{
+  std::ostringstream out;
+  this->ToString(_prefix,
+                 _includeDefaultElements,
+                 _includeDefaultAttributes,
+                 out);
   return out.str();
 }
 
 /////////////////////////////////////////////////
 void Element::ToString(const std::string &_prefix,
+                       bool _includeDefaultElements,
+                       bool _includeDefaultAttributes,
                        std::ostringstream &_out) const
 {
-  if (this->dataPtr->includeFilename.empty())
-  {
-    PrintValuesImpl(_prefix, _out);
-  }
-  else
-  {
-    _out << _prefix << "<include filename='"
-         << this->dataPtr->includeFilename << "'/>\n";
-  }
+  PrintValuesImpl(_prefix,
+                  _includeDefaultElements,
+                  _includeDefaultAttributes,
+                  _out);
 }
 
 /////////////////////////////////////////////////
@@ -942,18 +997,6 @@ void Element::Reset()
 void Element::AddElementDescription(ElementPtr _elem)
 {
   this->dataPtr->elementDescriptions.push_back(_elem);
-}
-
-/////////////////////////////////////////////////
-void Element::SetInclude(const std::string &_filename)
-{
-  this->dataPtr->includeFilename = _filename;
-}
-
-/////////////////////////////////////////////////
-std::string Element::GetInclude() const
-{
-  return this->dataPtr->includeFilename;
 }
 
 /////////////////////////////////////////////////