+++ Changelog.md
@@ -6,6 +6,51 @@
 
 ### libsdformat 11.X.X (202X-XX-XX)
 
+### libsdformat 11.2.2 (2021-07-01)
+
+1. Fix segfault when checking for required elements in joint
+    * [Pull request #610](https://github.com/ignitionrobotics/sdformat/pull/610)
+
+1. Add ValidateGraphs methods to Model/World
+    * [Pull request #601](https://github.com/ignitionrobotics/sdformat/pull/601)
+
+1. Making `PrintValues()` and `ToString()` able to not print default elements
+    * [Pull request #575](https://github.com/ignitionrobotics/sdformat/pull/575)
+
+### libsdformat 11.2.1 (2021-06-28)
+
+1. Fix ABI break on sdf11
+    * [Pull request #606](https://github.com/ignitionrobotics/sdformat/pull/606)
+
+1. Add triage, remove ticket templates
+    * [Pull request #608](https://github.com/ignitionrobotics/sdformat/pull/608)
+
+### libsdformat 11.2.0 (2021-06-23)
+
+1. Revert behavior of FilePath() for elements loaded from strings
+    * [Pull request #582](https://github.com/ignitionrobotics/sdformat/pull/582)
+
+1. Adding comment and modifications to element_tracing test regarding path sanitization
+    * [Pull request #578](https://github.com/ignitionrobotics/sdformat/pull/578)
+
+1. Allow to convert URDF color to SDF material tag
+    * [Pull request #526](https://github.com/ignitionrobotics/sdformat/pull/526)
+
+1. Encode XML path and line number in Element and add XML path setter Error.
+    * [Pull request #548](https://github.com/ignitionrobotics/sdformat/pull/548)
+
+1. Merge sdf 1.7 changes forward to 1.8
+    * [Pull request #570](https://github.com/ignitionrobotics/sdformat/pull/570)
+
+1. Add API for determining if an element was set by the user
+    * [Pull request #542](https://github.com/ignitionrobotics/sdformat/pull/542)
+
+1. Add scatter ratio parameter to Particle Emitter DOM
+    * [Pull request #547](https://github.com/ignitionrobotics/sdformat/pull/547)
+
+1. Fix bug where included URDFs are not parsed when there are no custom parsers
+    * [Pull request #562](https://github.com/ignitionrobotics/sdformat/pull/562)
+
 ### libsdformat 11.1.0 (2021-04-30)
 
 1. Clean up use of PROJECT_SOURCE_PATH in tests
@@ -170,6 +215,17 @@
     * [Pull request #460](https://github.com/osrf/sdformat/pull/460)
 
 ## libsdformat 10.X
+
+### libsdformat 10.5.0 (2021-05-17)
+
+1. Add scatter ratio parameter to Particle Emitter DOM.
+    + [Pull request 547](https://github.com/osrf/sdformat/pull/547)
+
+1. Methods for removing attributes from an element.
+    + [Pull request 555](https://github.com/osrf/sdformat/pull/555)
+
+1. Improve docs of collision bitmask.
+    + [Pull request 521](https://github.com/osrf/sdformat/pull/521)
 
 ### libsdformat 10.4.0 (2021-04-06)